client: move ProgressHelper to dedicated file
